Fix colors in the app
