--- app/src/components/color/categoricalColors.js
@@ -1,29 +1,0 @@
-/**
- * Copyright 2017-2020 Lawrence Livermore National Security, LLC and other
- * CallFlow Project Developers. See the top-level LICENSE file for details.
- * SPDX-License-Identifier: MIT
- */
-const categoricalColors = {
-	0: [0.31, 0.48, 0.65],
-	1: [0.94, 0.56, 0.22],
-	2: [0.87, 0.35, 0.36],
-	3: [0.47, 0.72, 0.70],
-	4: [0.36, 0.63, 0.32],
-	5: [0.93, 0.78, 0.33],
-	6: [0.69, 0.48, 0.63],
-	7: [0.99, 0.62, 0.66],
-	8: [0.61, 0.46, 0.38],
-	9: [0.73, 0.69, 0.67],
-	blue: [0.31, 0.48, 0.65],
-	orange: [0.94, 0.56, 0.22],
-	red: [0.87, 0.35, 0.36],
-	teal: [0.47, 0.72, 0.70],
-	green: [0.36, 0.63, 0.32],
-	yellow: [0.93, 0.78, 0.33],
-	purple: [0.69, 0.48, 0.63],
-	pink: "#610121",
-	brown: [0.61, 0.46, 0.38],
-	gray: [0.73, 0.69, 0.67]
-};
-
-export default categoricalColors;
