+++ app/src/components/colors.js
@@ -1,0 +1,23 @@
+/**
+ * Copyright 2017-2020 Lawrence Livermore National Security, LLC and other
+ * CallFlow Project Developers. See the top-level LICENSE file for details.
+ * SPDX-License-Identifier: MIT
+ */
+import tpl from "../../html/scatterplot/index.html";
+
+export default {
+	template: tpl,
+	name: "Colors",
+	props: [],
+	data: () => ({
+
+	}),
+	mounted() {
+
+	},
+	methods: {
+		init() {
+            this.color = 
+		},
+	}
+};
