+++ app/src/components/ensembleSupergraph/colormap.js
@@ -7,6 +7,7 @@ import * as d3 from "d3";
 import "d3-selection-multi";
 import * as utils from "../utils";
 import EventHandler from "../EventHandler";
+import { render } from "dagre-d3";
 
 export default {
 	template: "<g :id=\"id\"></g>",
@@ -17,9 +18,8 @@ export default {
 
 	data: () => ({
 		transitionDuration: 1000,
-		width: 200,
+		width: 230,
 		height: 20,
-		colorScaleHeight: 30,
 		colorMin: 0,
 		colorMax: 0,
 		offset: 30,
@@ -27,12 +27,10 @@ export default {
 			bottom: 30,
 			right: 400,
 		},
-		id: ""
+		id: "ensemble-colormap"
 	}),
 
 	mounted() {
-		this.id = "ensemble-colormap";
-
 		EventHandler.$on("show_target_auxiliary", (data) => {
 			this.init();
 		});
@@ -43,41 +41,67 @@ export default {
 			this.colorMin = this.$store.color.colorMin;
 			this.colorMax = this.$store.color.colorMax;
 			this.innerHTMLText = [this.$store.colorMin, this.$store.colorMax];
-			this.colorMap = this.$store.color.colorMap;
-			this.colorPoint = this.$store.colorPoint;
 
-			this.parentID = this.$parent.id;
 			this.containerWidth = this.$store.viewWidth / 2;
 			this.containerHeight = this.$store.viewHeight - this.$parent.margin.top - this.$parent.margin.bottom;
 
-			this.clearColorMap();
-			this.svg = d3.select("#" + this.parentID)
+			this.svg = d3.select("#" + this.$parent.id)
 				.append("g")
 				.attrs({
 					"id": "dist-colormap",
 				});
-
-			this.clearTargetLegends();
-			if (this.$store.showTarget == true) {
-				this.drawTargetLegend();
-			}
-			if(this.$store.selectedMode == 'Ensemble'){
-				this.clearEnsembleLegends();
-				this.drawEnsembleLegend();	
-			}
-			this.drawMeanColorMap();
-			this.drawMeanText();
+			
+			render()
 		},
 
-		drawEnsembleLegend() {
+		render() {
+			this._legends()
+			this._color_map()
+		},
+
+		_legends() {
+			this.clearTargetLegends();
+			if (this.$store.showTarget == true) {
+				this.drawLegend("Target run", this.containerWidth - this.padding.right, this.containerHeight - 4 * this.padding.bottom, this.$store.color.target)
+			}
+			if (this.$store.selectedMode == 'Ensemble') {
+				this.clearEnsembleLegends();
+				this.drawLegend("Ensemble of runs", this.containerWidth - this.padding.right, this.containerHeight - 3 * this.padding.bottom, this.$store.color.ensemble)
+			}
+		},
+
+		_color_map() {
+			this.clearColorMap();
+			if (this.$store.mode == "mean") {
+				let text = ""
+				if (this.$store.selectedMetric == "Exclusive") {
+					text = "Exc. Runtime colormap"
+				}
+				else if (this.$store.selectedMetric == "Inclusive") {
+					text = "Inc. Runtime colormap"
+				}
+				this.drawColorMap(text, this.containerWidth - this.padding.right, this.containerHeight - this.padding.bottom);
+			}
+			else if (this.$store.mode == "mean-gradients") {
+				this.drawColorMap("Distribution colormap", this.containerWidth - this.padding.right, this.containerHeight - 2 * this.padding.bottom)
+			}
+			else if (this.$store.mode == "mean-diff") {
+				this.drawColorMap("Mean Difference colormap", this.containerWidth - this.padding.right, this.containerHeight - 2 * this.padding.bottom)
+			}
+			else if (this.$store.mode == "rank-diff") {
+				this.drawColorMap("Rank Difference colormap", this.containerWidth - this.padding.right, this.containerHeight - 2 * this.padding.bottom)
+			}
+		},
+
+		drawLegend(text, x, y, color) {
 			this.svg.append("circle")
 				.attrs({
 					"r": 10,
 					"cx": 10,
 					"cy": 10,
 					"class": "ensemble-circle-legend",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 3 * this.padding.bottom})`,
-					"fill": this.$store.color.ensemble
+					"transform": `translate(${x}, ${y})`,
+					"fill": color
 				});
 
 			this.svg.append("text")
@@ -85,137 +109,28 @@ export default {
 					"x": 30,
 					"y": 15,
 					"class": "ensemble-circle-legend-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 3 * this.padding.bottom})`,
+					"transform": `translate(${x}, ${y})`,
 				})
-				.text("Ensemble of runs")
-				.style("font-size", 14)
-				.style("fill", "#444444");
-
-
-		},
-
-		drawTargetLegend() {
-			this.svg.append("circle")
-				.attrs({
-					"r": 10,
-					"cx": 10,
-					"cy": 10,
-					"class": "target-circle-legend",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 4 * this.padding.bottom})`,
-					"fill": this.$store.color.target
-				});
-
-			this.svg.append("text")
-				.attrs({
-					"x": 30,
-					"y": 15,
-					"class": "target-circle-legend-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 4 * this.padding.bottom})`,
-				})
-				.text("Target run")
+				.text(text)
 				.style("font-size", 14)
 				.style("fill", "#444444");
 		},
 
-		drawMeanColorMap() {
-			this.color = this.$store.color;
-			if (this.color.option == "Module") {
-				console.log("TODO");
-			} else {
-				let splits = this.$store.colorPoint;
-				let color = this.color.getScale(this.color.option);
-
-				for (let i = 0; i < splits; i += 1) {
-					let splitColor = this.colorMin + ((i * this.colorMax) / (splits));
-					this.svg.append("rect")
-						.attrs({
-							"width": this.width / splits,
-							"height": this.height,
-							"x": i * (this.width / splits),
-							"class": "dist-colormap-rect-metric",
-							"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - this.padding.bottom})`,
-							"fill": color(splitColor)
-						});
-				}
+		drawColorMap(text, x, y) {
+			let splits = this.$store.selectedColorPoint;
+			for (let i = 0; i < splits; i += 1) {
+				let splitColor = this.colorMin + ((i * this.colorMax) / (splits));
+				this.svg.append("rect")
+					.attrs({
+						"width": this.width / splits,
+						"height": this.height,
+						"x": i * (this.width / splits),
+						"class": "dist-colormap-rect-metric",
+						"transform": `translate(${x}, ${y})`,
+						"fill": this.$store.color.getColorByValue(splitColor)
+					});
 			}
-		},
 
-		drawRankDiffColorMap() {
-			this.color = this.$store.rankDiffColor;
-
-			if (this.color.option == "Module") {
-				console.log("TODO");
-			} else {
-				let splits = this.$store.colorPoint;
-				let color = this.color.getScale("RankDiff");
-				for (let i = 0; i < splits; i += 1) {
-					let splitColor = this.colorMin + ((i * this.colorMax) / (splits));
-					this.svg.append("rect")
-						.attrs({
-							"width": this.width / splits,
-							"height": this.height,
-							"x": i * (this.width / splits),
-							"class": "dist-colormap-rect",
-							"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-							"fill": color(splitColor)
-						});
-				}
-			}
-		},
-
-		drawMeanDiffColorMap(splits, colorscale) {
-			this.color = this.$store.meanDiffColor;
-
-			if (this.color.option == "Module") {
-				console.log("TODO");
-			} else {
-				// let splits = this.$store.colorPoint
-				let splits = 9;
-				let color = this.color.getScale("MeanDiff");
-				let domain = color.domain();
-				let colorMin = domain[0];
-				let colorMax = domain[1];
-				let dcolor = (colorMax - colorMin) / (splits - 1);
-				for (let i = 0; i < splits; i += 1) {
-					let splitColor = colorMin + dcolor * (splits - 1 - i);
-
-					this.svg.append("rect")
-						.attrs({
-							"width": this.width / splits,
-							"height": this.height,
-							"x": i * (this.width / splits),
-							"class": "dist-colormap-rect",
-							"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-							"fill": color(splitColor)
-						});
-				}
-			}
-		},
-
-		drawBinColorMap() {
-			this.color = this.$store.binColor;
-
-			if (this.color.option == "Module") {
-				console.log("TODO");
-			} else {
-				let splits = this.$store.colorPoint;
-				let color = this.color.getScale("Bin");
-				for (let i = 0; i < splits; i += 1) {
-					let splitColor = this.colorMin + ((i * this.colorMax) / (splits));
-					this.svg.append("rect")
-						.attrs({
-							"width": this.width / splits,
-							"height": this.height,
-							"x": i * (this.width / splits),
-							"class": "dist-colormap-rect",
-							"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-							"fill": color(splitColor)
-						});
-				}
-			}
-		},
-
-		drawMeanText() {
 			this.svg.append("text")
 				.style("fill", "black")
 				.style("font-size", "14px")
@@ -225,9 +140,9 @@ export default {
 					"x": -125,
 					"text-anchor": "middle",
 					"class": "dist-colormap-text-metric",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - this.padding.bottom})`,
+					"transform": `translate(${x}, ${y})`,
 				})
-				.text("Exc. Runtime colormap");
+				.text(text);
 
 			// draw the element
 			this.svg.append("text")
@@ -239,9 +154,9 @@ export default {
 					"y": 10,
 					"x": -40,
 					"class": "dist-colormap-text-metric",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - this.padding.bottom})`,
+					"transform": `translate(${x}, ${y})`,
 				})
-				.text(utils.formatRuntimeWithUnits(this.colorMin));
+				.text(utils.formatRuntimeWithUnits(this.$store.color.colorMin));
 
 			this.svg.append("text")
 				.style("fill", "black")
@@ -252,124 +167,9 @@ export default {
 					"y": 10,
 					"x": 30,
 					"class": "dist-colormap-text-metric",
-					"transform": `translate(${this.containerWidth - this.padding.right + this.width + this.offset}, ${this.containerHeight - this.padding.bottom})`,
+					"transform": `translate(${x + this.width}, ${y})`,
 				})
-				.text(utils.formatRuntimeWithUnits(this.colorMax));
-
-		},
-
-		drawRankDiffText() {
-			// draw the element
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": -40,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text(this.colorMin);
-
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": 30,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right + this.width}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text(this.colorMax);
-
-		},
-
-		drawMeanDiffText() {
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": -160,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text("Mean Difference colormap");
-
-			// draw the element
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": -40,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text(utils.formatRuntimeWithUnits(this.colorMin));
-
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": 30,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right + this.width + this.offset}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text(utils.formatRuntimeWithUnits(this.colorMax));
-		},
-
-		drawBinText() {
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": -120,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text("Distribution colormap");
-
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": -30,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text(this.colorMin);
-
-			this.svg.append("text")
-				.style("fill", "black")
-				.style("font-size", "14px")
-				.attrs({
-					"dy": ".35em",
-					"y": 10,
-					"x": 10,
-					"text-anchor": "middle",
-					"class": "dist-colormap-text",
-					"transform": `translate(${this.containerWidth - this.padding.right + this.width + this.offset}, ${this.containerHeight - 2 * this.padding.bottom})`,
-				})
-				.text(this.colorMax);
-
+				.text(utils.formatRuntimeWithUnits(this.$store.color.colorMax));
 		},
 
 		clearColorMap() {
@@ -396,48 +196,15 @@ export default {
 			d3.selectAll(".ensemble-circle-legend-text").remove();
 		},
 
-		update(mode, data) {
-			this.clear();
-			let rank_min = 0;
-			let rank_max = 0;
-			let mean_min = 0;
-			let mean_max = 0;
-			let mean_diff_min = 0;
-			for (let i = 0; i < data.length; i += 1) {
-				rank_min = Math.min(rank_min, data[i]["hist"]["y_min"]);
-				rank_max = Math.max(rank_max, data[i]["hist"]["y_max"]);
-				mean_min = Math.min(mean_min, data[i]["hist"]["x_min"]);
-				mean_max = Math.max(mean_max, data[i]["hist"]["x_max"]);
-				mean_diff_min = Math.min(this.colorMin, data[i]["hist"]["mean_diff"]);
-			}
-			if (mode == "rankDiff") {
-				this.colorMin = rank_min;
-				this.colorMax = rank_max;
-				this.drawRankDiffColorMap();
-				this.drawRankDiffText();
-			}
-			else if (mode == "meanDiff") {
-				this.colorMin = mean_diff_min;
-				this.colorMax = mean_max;
-				this.drawMeanDiffColorMap();
-				this.drawMeanDiffText();
-			}
-		},
-
-		updateWithMinMax(mode, min, max) {
+		update(mode, min, max) {
 			this.clear();
 			this.colorMin = min;
 			this.colorMax = max;
-			if (mode == "bin") {
-				this.drawBinColorMap();
-				this.drawBinText();
-			}
-			if (mode == "meanDiff") {
+			if (mode == "mean-difference" || mode == "rank-difference") {
 				this.colorMin = -1 * Math.max(Math.abs(min), Math.abs(max));
 				this.colorMax = 1 * Math.max(Math.abs(min), Math.abs(max));
-				this.drawMeanDiffColorMap();
-				this.drawMeanDiffText();
 			}
+			this.render()
 		}
 	}
 };
